relocate code of interpretercore gc
