+++ paddle/fluid/framework/new_executor/interpretercore.cc
@@ -16,15 +16,12 @@
 #include <unordered_set>
 #include "paddle/fluid/framework/details/nan_inf_utils.h"
 #include "paddle/fluid/framework/details/share_tensor_buffer_functor.h"
-#include "paddle/fluid/framework/new_executor/interpretercore_event_garbage_collector.h"
+#include "paddle/fluid/framework/new_executor/garbage_collector/interpretercore_event_garbage_collector.h"
+#include "paddle/fluid/framework/new_executor/garbage_collector/interpretercore_fast_garbage_collector.h"
 #include "paddle/fluid/framework/new_executor/interpretercore_util.h"
 #include "paddle/fluid/framework/operator.h"
 #include "paddle/fluid/platform/os_info.h"
 #include "paddle/fluid/platform/profiler/event_tracing.h"
-
-#if defined(PADDLE_WITH_CUDA) || defined(PADDLE_WITH_HIP)
-#include "paddle/fluid/framework/new_executor/interpretercore_fast_garbage_collector.h"
-#endif
 
 PADDLE_DEFINE_EXPORTED_bool(new_executor_use_inplace, true,
                             "Use inplace in new executor");
