--- paddle/fluid/framework/new_executor/interpretercore_fast_garbage_collector.h
@@ -1,30 +1,0 @@
-// Copyright (c) 2021 PaddlePaddle Authors. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-#pragma once
-
-#include "paddle/fluid/framework/new_executor/interpretercore_garbage_collector.h"
-
-namespace paddle {
-namespace framework {
-
-class InterpreterCoreFastGarbageCollector
-    : public InterpreterCoreGarbageCollector {
- public:
-  virtual void Add(Variable* var) override;
-
- private:
-  void Add(Garbage garbage);
-};
-}  // namespace framework
-}  // namespace paddle
